changed file name, refactor and nuke11 support
